[UPDATE] socket STOMP 코드 주석처리
